news_service/requirements.txt
Django==2.2.19
djangorestframework==3.12.4
flake8==6.0.0
flake8-plugin-utils==1.3.3
gunicorn==20.0.4
isort==5.12.0
mccabe==0.7.0
psycopg2-binary==2.8.6
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
